+++ packages/plugins/GeoLocation/src/store/getters.ts
@@ -8,6 +8,11 @@ const getters: PolarGetterTree<GeoLocationState, GeoLocationGetters> = {
   boundaryLayerId: (_, __, ___, rootGetters): string | undefined => {
     return rootGetters?.configuration?.geoLocation?.boundaryLayerId
   },
+  boundaryOnError: (_, __, ___, rootGetters) => {
+    return (
+      rootGetters.configuration?.geoLocation?.boundaryOnError || 'permissive'
+    )
+  },
   configuredEpsg: (_, __, ___, rootGetters): string | undefined => {
     return rootGetters?.configuration?.epsg
   },
@@ -22,6 +27,9 @@ const getters: PolarGetterTree<GeoLocationState, GeoLocationGetters> = {
       return keepCentered
     }
     return false
+  },
+  showTooltip: (_, __, ___, rootGetters): boolean => {
+    return Boolean(rootGetters.configuration?.geoLocation?.showTooltip)
   },
   toastAction: (_, __, ___, rootGetters): string | undefined => {
     return rootGetters?.configuration?.geoLocation?.toastAction