+++ packages/types/custom/core.ts
@@ -228,6 +228,8 @@ export interface GfiLayerConfiguration {
   format?: string
 }
 
+export type BoundaryOnError = 'strict' | 'permissive'
+
 export interface LayerBoundPluginOptions extends PluginOptions {
   /**
    * If set, feature will only be applicable within the layer's features.
@@ -235,6 +237,7 @@ export interface LayerBoundPluginOptions extends PluginOptions {
    * selecting unfit coordinates.
    */
   boundaryLayerId?: string
+  boundaryOnError?: BoundaryOnError
   /**
    * Used if boundaryLayer does not contain the plugin information to inform
    * the user that something could not be set/updated.
@@ -251,6 +254,7 @@ export interface GeoLocationConfiguration extends LayerBoundPluginOptions {
   checkLocationInitially: boolean
   /** whether to keep center on user or allow movement after first zoom to */
   keepCentered: boolean
+  showTooltip?: boolean
   /**
    * Limits the viewable GFIs per layer by this number. The first n elements
    * are chosen arbitrarily. Useful if you e.g. just want one result, or to
@@ -345,12 +349,12 @@ export interface PinsConfiguration extends LayerBoundPluginOptions {
   appearOnClick: AppearOnClick
   /** Path in store from where coordinates can be retrieved from. */
   coordinateSource: string
-  /** The zoom level to zoom to when a pin is added to the map. */
   initial?: InitialPin
   /** If the pin should be movable; defaults to false. */
   movable?: boolean | MovablePin
   /** Pin styling */
   style?: PinStyle
+  /** The zoom level to zoom to when a pin is added to the map. */
   toZoomLevel: number
 }
 
@@ -531,11 +535,13 @@ export interface PolarError {
 
 export interface CoreState {
   map: number
+  center: [number, number] | null
   clientHeight: number
   clientWidth: number
   components: number
   configuration: MapConfig
   errors: PolarError[]
+  plugin: object
 }
 
 export interface CoreGetters {